+++ ext-src/lib/commands/InvokeLocal.ts
@@ -1,6 +1,5 @@
-import * as _ from 'lodash';
 import * as path from 'path';
-import { ExtensionContext, Uri, window } from 'vscode';
+import { ExtensionContext } from 'vscode';
 import { CommandBase } from '../CommandBase';
 import { Serverless } from '../Serverless';
 import { NodeKind, ServerlessNode } from '../ServerlessNode';
@@ -22,16 +21,14 @@ export class InvokeLocal extends CommandBase {
 
 		return new Promise(async (resolve, reject) => {
 			try {
-				const result = await CommandBase.getConfig();
-				const functionName: string = node.name
-				const path: string = `${node.documentRoot}/test/${functionName}.json`;
+				const config = await CommandBase.getConfig();
 
 				resolve(
 					Serverless.invoke('invoke local', {
 						'cwd': node.documentRoot,
-						'function': functionName,
-						'path': path,
-						...result,
+						'function': node.name,
+						'path': path.join(node.documentRoot, config.testFolderPath, `${node.name}.json`),
+						...config,
 						'log': false
 					})
 				);